pyproject.toml
[tool.poetry]
name = "inventorization_system"
version = "0.1.0"
description = "My diploma project"
authors = ["Heorhii <george.sokolovsky2001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.5.1"
"backports.zoneinfo" = "0.2.1"
coverage = "6.3.2"
Django = "4.0.4"
djangorestframework = "3.13.1"
flake8 = "4.0.1"
mccabe = "0.6.1"
pika = "1.2.1"
pika-stubs = "0.1.3"
pycodestyle = "2.8.0"
pyflakes = "2.4.0"
pytz = "2022.1"
sqlparse = "0.4.2"
drf-nested-routers = "^0.93.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
